Allow support for an external avatar image
By defaul the summary.html file assumes that you'll be using a local
resource, and therefore constructs a relative url. These changes will
check for the presence of `http` or `//` and change the `img` element
to
use an absolute path. This allows you to use an external avatar image,
such as Gravatar.